onUnityAdsFinish function implementation diverged so much that I separated them from BaseAdFullScreen
